requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.6
django-bootstrap-icons==0.8.1
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-mathfilters==1.0.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.4
kombu==5.2.4
numpy==1.23.1
packaging==21.3
pandas==1.4.3
prompt-toolkit==3.0.31
psycopg==3.0.15
psycopg2-binary==2.8.4
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
